rename namespaces to issuer namespaces
